pyproject.toml
[tool.poetry]
name = "slice-qtc"
version = "0.1.0"
description = "A package to communicate with a VESCENT Slice-QTC temperature controller."
authors = ["Julian Stuermer stuermer@lsw.uni-heidelberg.de"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
coverage = "^6.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
